Bug fix double precision in MET-file
git-svn-id: https://svn.oss.deltares.nl/repos/imod/trunk@2971 8a8853d3-af7d-47ec-b359-4bd2096ba547

+++ MODFLOW/iMODUTL/rf2mf_mod_conv.f90
@@ -384,7 +384,7 @@
          character(len=maxlen) :: keyword
          integer :: type = imetu
          integer :: ival
-         real    :: rval
+         real(KIND=8) :: rval
          character(len=maxlen) :: cval = ''
          type(tTime) :: time
       end type tKw
